Added example of DialogView using
