+++ chat/src/main/res/layout/item_dialog.xml
@@ -87,6 +87,7 @@
             android:layout_marginTop="7dp"
             android:ellipsize="end"
             android:maxLines="1"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toEndOf="@+id/lastAuthorAvatar"
             app:layout_constraintTop_toBottomOf="@+id/dialogName" />
